% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clrSIEVE.R
\name{clrSIEVE}
\alias{clrSIEVE}
\title{Simultaneous Differential Expression, Variability and Skewness Analysis Using RNA-Seq Data}
\usage{
clrSIEVE(
  clrSeq_result = NULL,
  alpha_level = 0.05,
  order_DE = FALSE,
  order_LFC = FALSE,
  order_DS = FALSE,
  order_sieve = FALSE
)
}
\arguments{
\item{clrSeq_result}{The result of clrSeq() function.}

\item{alpha_level}{The adjusted p-value cutoff used for flagging genes that show significant
differential expression, variability and skewness.}

\item{order_DE}{Logical string. "FALSE" for no ordering; "TRUE" for ordering by the value of \code{DE}.}

\item{order_LFC}{Logical string. "FALSE" for no ordering; "TRUE" for ordering by the value of \code{LFC}.}

\item{order_DS}{Logical string. "FALSE" for no ordering; "TRUE" for ordering by the value of \code{DS}.}

\item{order_sieve}{Character/logical string specifying the order method. Possibilities are "DE" for
the value of \code{DE}, "LFC" for log2 fold change of variability,
"DS" for the value of \code{DS} or "FALSE" for no ordering.}
}
\value{
\code{clrDE_test}, \code{clrDV_test} and \code{clrDS_test} contain the
result of the DE, DV and DS tests, respectively.
\code{clrSIEVE_tests} contains the result of all three tests.
\item{clrDE_test}{A data.frame contating the results of differential expression test.}
\item{clrDV_test}{A data.frame contating the results of differential variability test.}
\item{clrDS_test}{A data.frame contating the results of differential skewness test.}
\item{clrSIEVE_tests}{A data.frame contating the results of differential expression, variability and skewness tests.}
\item{DE}{The difference of mean (\code{mu}) between group 2 and group 1 (\code{mu2} \code{-} \code{mu1}).}
\item{se_DE}{The standard error of \code{DE}.}
\item{z_DE}{The observed Wald statistic of \code{DE}.}
\item{pval_DE}{The unadjusted p-value of Wald test of \code{DE}.}
\item{adj_pval_DE}{The p-value of the Wald test of \code{DE}, adjusted using the Benjamini-Yekutieli procedure.}
\item{mu1}{The maximum likelihood estimate of the mean parameter for group 1.}
\item{mu2}{The maximum likelihood estimate of the mean parameter for group 2.}
\item{de_indicator}{1: DE gene; 0: non-DE gene.}
\item{SD_ratio}{The ratio of standard deviation (\code{sigma}) between group 2 and group 1 (\code{sigma2}\code{/}\code{sigma1}). }
\item{LFC}{\code{log2} fold change: \code{log2}(\code{sigma2}\code{/}\code{sigma1}).}
\item{DV}{The difference of standard deviation (\code{sigma}) between group 2 and group 1 (\code{sigma2} \code{-} \code{sigma1}).}
\item{se_DV}{The standard error of \code{DV}.}
\item{z_DV}{The observed Wald statistic of \code{DV}.}
\item{pval_DV}{The unadjusted p-value of Wald test of \code{DV}.}
\item{adj_pval_DV}{The p-value of the Wald test of \code{DV}, adjusted using the Benjamini-Yekutieli procedure.}
\item{sigma1}{The maximum likelihood estimate of the standard deviation parameter for group 1.}
\item{sigma2}{The maximum likelihood estimate of the standard deviation parameter for group 2.}
\item{dv_indicator}{1: DV gene; 0: non-DV gene.}
\item{DS}{The difference of skewness parameter (\code{gamma}) between group 2 and group 1 (\code{gamma2} \code{-} \code{gamma1}.).}
\item{se_DS}{The standard error of \code{DS}.}
\item{z_DS}{The observed Wald statistic of \code{DS}.}
\item{pval_DS}{The unadjusted p-value of Wald test of \code{DS}.}
\item{adj_pval_DS}{The p-value of the Wald test of \code{DS} adjusted using the Benjamini-Yekutieli procedure.}
\item{gamma1}{The maxiimum likelihood estimate of the skewness parameter for group 1.}
\item{gamma2}{The maxiimum likelihood estimate of the skewness parameter for group 2.}
\item{ds_indicator}{1: DS gene; 0: non-DS gene.}
}
\description{
Model CLR-transformed RNA-Seq data using the skew-normal
distribution, and then conduct statistical tests for finding
genes/transcripts with differential expression, variability
and skewness using the Wald test.
}
\examples{
library(SIEVEseq)
data(clrCounts3) # first 50 genes (gene1 to gene50) are DE genes
groups <- c(rep(0, 200), rep(1, 200))
clrSeq_result3 <- clrSeq(clrCounts3[46:100, ], group = groups) # DE dataset
clrSIEVE_result3 <- clrSIEVE(clrSeq_result = clrSeq_result3,
                             alpha_level = 0.05,
                             order_DE = FALSE,
                             order_LFC = FALSE,
                             order_DS = FALSE,
                             order_sieve = FALSE)
clrDE_test3 <- clrSIEVE_result3$clrDE_test # DE test
head(clrDE_test3, 5)
clrDS_test3 <- clrSIEVE_result3$clrDS_test # DS test
clrDS_test3[clrDS_test3$adj_pval_DS < 0.05, ]
clrSIEVE_tests3 <- clrSIEVE_result3$clrSIEVE_tests # Sieve DE, DV and DS genes
head(clrSIEVE_tests3, 5)
tail(clrSIEVE_tests3, 5)

}
\references{
{
Azzalini, A. (1985). \emph{A class of distributions which
includes the normal ones}. \emph{Scandinavian Journal of Statistics} \bold{12}(2),
171--178, JSTOR

Azzalini, A. and Capitanio, A. (2014).
\emph{The Skew-Normal and Related Families}.
Cambridge University Press, IMS Monographs series.

Azzalini, A. and Arellano-Valle, R. B. (2013).
Maximum penalized likelihood estimation for skew-normal and skew-\emph{t}
distributions. \emph{Journal of Statistical Planning and Inference} \bold{143}, 419--433.

Azzalini, A. (2022). \emph{The R package \bold{sn}: The skew-normal
and related distribution such as the skew-t and the SUN (version 2.0.2)}.
Universit\`a degli Studi di Padova, Italia.
Home page: \url{https://cran.r-project.org/package=sn}.

Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
Chapman & Hall, London.
}
}
\seealso{
\code{\link[=clrSeq]{clrSeq()}} for the results of parameters estimation; alternatively, \code{\link[=clrDE]{clrDE()}} only provides DE test and \code{\link[=clrDV]{clrDV()}} only provides DV test.
}
